ai_assistant/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
python-dotenv
pymupdf
langchain
langchain-google-genai
langchain-chroma
chromadb
streamlit
requests
